Kewords RAG/requirements.txt
# requirements.txt
# Core dependencies for Google Ads Keyword RAG System

# Google Ads API
google-ads==22.1.0

# OpenAI API
openai>=1.0.0

# Vector embeddings and similarity search
sentence-transformers==2.2.2
faiss-cpu==1.7.4
# faiss-gpu==1.7.4  # Uncomment for GPU support

# Data processing
pandas==2.1.4
numpy==1.24.3

# Async HTTP requests
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0

# Optional: For enhanced text processing
nltk==3.8.1
spacy==3.7.2

# Optional: For data visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# Optional: For web interface
flask==3.0.0
streamlit==1.29.0

# Development dependencies
pytest==7.4.3
black==23.12.0
flake8==6.1.0
mypy==1.8.0

# Jupyter notebook support (optional)
jupyter==1.0.0
ipykernel==6.27.1